Deep diff on a single unit.
