+++ maintenance.js
@@ -7,8 +7,7 @@ const util = require('util');
 const fs = require('fs');
 const models       = require('./db/models');
 const dbinit       = require('./db/init');
-const assert = require('assert').strict;
-const deepd        = require('deep-diff');
+const diff = require('./units/diff');
 // load local config and check if is ok (testing db)
 const localconfig = dbinit.init();
 const hasha = require('hasha');
@@ -50,46 +49,6 @@ function regenerateMaps (maps) {
     }
 }
 
-function processDeepDiff(hists) {
-    if (hists.length) {
-        try {
-            let recordA = hists.shift();
-            // console.log(recordA);
-            // process.exit(1);
-            let recordB = hists[0];  // keep for next check
-            // console.log(recordA);
-            // process.exit(1);
-            util.log("Show diff between id %d and %d", recordA.get('id'), recordB.get('id'));
-            let differences = deepd.diff(JSON.parse(recordA.json), JSON.parse(recordB.json));
-            console.log(differences);
-            // display diff
-            let recArr = [recordA, recordB];
-            for (rec of recArr) {
-                let nomeFile = util.format("local/record_%d.json", rec.get('id'));
-                fs.writeFile(nomeFile, JSON.stringify(JSON.parse(rec.json), 2, '\t'), (err) => {
-                    // throws an error, you could also catch it here
-                    if (err) throw err;
-
-                    // success case, the file was saved
-                    console.log('saved!');
-                    processDeepDiff(hists);
-                });
-            }
-        }
-        catch (e) {
-            util.log("Cannot get diff");
-        }
-    }
-    else {
-        util.log("deep diff ok");
-        process.exit(0);
-        // exit without errors
-        // process.exit(0);
-
-
-    }
-}
-
 if (program.regeneratepreviews)  {
     // Used for landing page, 3 elements per load, offset passed by url
     let dbMeta = new db.Database(localconfig.database);
@@ -112,32 +71,38 @@ if (program.regeneratepreviews)  {
     });
 }
 if (program.testdiff)  {
-  let dbMeta = new db.Database(localconfig.database);
-  const Map = dbMeta.db.define('map', models.Map);
-  const History = dbMeta.db.define('history', models.History);
-  Map.hasMany(History); // 1 : N
-  History.belongsTo(Map);  // new property named "map" for each record
-  // specify map.id by command line
-  var historyWhere = {mapId:  parseInt(program.testdiff)};
-  historyWhere['diff'] = true;
-  History.findAll({
-    where: historyWhere,
-    include: [{
-        model: Map,
-        where: {
-          published: true
+    let dbMeta = new db.Database(localconfig.database);
+    const Map = dbMeta.db.define('map', models.Map);
+    const History = dbMeta.db.define('history', models.History);
+    Map.hasMany(History); // 1 : N
+    History.belongsTo(Map);  // new property named "map" for each record
+    // specify map.id by command line
+    var historyWhere = {mapId:  parseInt(program.testdiff)};
+    historyWhere['diff'] = true;
+    History.findAll({
+      where: historyWhere,
+      include: [{
+          model: Map,
+          where: {
+            published: true
+          }
         }
-      }
-    ],
-    order: [
-      ['createdAt', 'ASC']
-    ],
-    // offset: ???,
-    limit: 20
-  }).then(hists => {
-    processDeepDiff(hists);
-    // for (hist of hists) {
-    //   util.log('ok %d', hist.id);
-    // }
-  });
+      ],
+      order: [
+        ['createdAt', 'ASC']
+      ],
+      // offset: ???,
+      limit: 20
+    }).then(hists => {
+      diff.processDeepDiff(hists, function (results) {
+          for (r of results) {
+              util.log('----------------------------------------------------');
+              console.log(r.length);
+          }
+          process.exit(0)
+      });
+      // for (hist of hists) {
+      //   util.log('ok %d', hist.id);
+      // }
+    });
 }
